Added classificator

+++ src/main/java/Logic/Features/IFeature.java
@@ -4,6 +4,6 @@ import java.util.List;
 import java.util.Map;
 
 public interface IFeature {
-    int count(List<String> listOfWords);
+    double count(List<String> listOfWords);
     Map<String, Integer> count(Map<List<String>, String> data);
 }
